Fixed a bug related to default values in QtBaseViewerStereo, so stereo anaglyph mode works again now.
Added OpenFlipperQSettings which wraps the value() function to write an error message to console if that function is used without a default value. This is meant as a reminder to developers to properly initialize their settings and specify default values such that hopefully less bugs like the one above occur from now on.

Added default values for many calls to OpenFlipperSettings().value() - if you find any that I have overlooked, please add default values there, too!

git-svn-id: http://www.openflipper.org/svnrepo/OpenFlipper/branches/Free@10424 383ad7c9-94d9-4d36-a494-682f7c89f535